test: fix redis unit tests
